src/requirements.txt
openai==0.28
requests==2.26.0
streamlit
pinecone-client